fix(bots/discord): fix freeze on prod builds
